+++ packages/web/app/new-team/page.tsx
@@ -1,7 +1,7 @@
-import NewTeamForm from "@/components/new-team-form";
 import { Session } from "@tableland/studio-api";
 import { cookies } from "next/headers";
 import { notFound } from "next/navigation";
+import NewTeamForm from "./_components/new-team-form";
 
 export default async function NewProject() {
   const session = await Session.fromCookies(cookies());
